safari-600.7-branch/LayoutTests/inspector-obsolete/timeline/timeline-paint-expected.txt

Tests the Timeline API instrumentation of a paint event


Paint Properties:
{
    children : <object>
    counters : <object>
    data : {
        clip : <object>
    }
    endTime : <number>
    frameId : <string>
    stackTrace : <object>
    startTime : <number>
    type : "Paint"
    usedHeapSize : <number>
}
